+++ docs/changelog.rst
@@ -28,3 +28,8 @@ Changed
 
 - Updated publish workflow track to 16/edge.
 
+Added
+~~~~~
+
+- Terraform module to deploy Frappe HRMS with MariaDB and Redis.
+